rc: add BB-BONELT-HDMIN virtual cape
On the BealgeBone Black, the on-board HDMI acts like a "virtual"
cape; it uses pins on the expansion header and can conflict with
other capes.

This commit moves the tda19988 service start-up from rc to it's
own script so that it can be easily disabled automatically when
an LCD cape is present and enabled the rest of the time.

Change-Id: I564463dd2fc34eab44dfe65a61038d5c89407730